moved: 관리자와 멤버 로직을 구분짓기 위해 관리자용, 멤버용에 따른 api와 로직들 구조변경
