+++ surbear-api/src/main/java/com/surbear/member/service/MemberService.java
@@ -13,8 +13,6 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.Optional;
-
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
 @Service
@@ -53,17 +51,12 @@ public class MemberService {
         return true;
     }
 
-    public Member getMemberInfoByNickname(String nickname) {
-        return checkNicknameExists(nickname);
-    }
-
     public Member getMemberInfoByToken(Long memberId) {
         MemberEntity memberEntity = repository.findById(memberId).get();
         Member member = mapper.toModel(memberEntity);
         return member;
 
     }
-
 
     public boolean checkUserIdAndEmailExists(String userId, String email) {
         Member member = repository.findByUserIdAndEmail(userId, email);
@@ -92,14 +85,6 @@ public class MemberService {
 
     private Member checkUserIdExists(String userId) {
         Member member = repository.findByUserId(userId);
-        if (member == null) {
-            throw new ProcessErrorCodeException(BasicErrorCode.USER_NOT_FOUND);
-        }
-        return member;
-    }
-
-    private Member checkNicknameExists(String nickname) {
-        Member member = repository.findByNicknameAndDeletedIsFalse(nickname);
         if (member == null) {
             throw new ProcessErrorCodeException(BasicErrorCode.USER_NOT_FOUND);
         }
