+++ surbear-api/src/main/java/com/surbear/role/controller/RoleController.java
@@ -1,10 +1,10 @@
 package com.surbear.role.controller;
 
 
+import com.surbear.member.model.Member;
 import com.surbear.role.controller.dto.AdminListResponse;
 import com.surbear.role.service.FacadeRoleService;
 import com.surbear.role.service.RoleService;
-import com.surbear.survey.dto.survey.history.ParticipatedSurvey;
 import com.surbear.survey.question.model.Survey;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
@@ -41,15 +41,16 @@ public class RoleController {
         return roleService.getAdminList();
     }
 
-    @Operation(summary = "유저 참여 설문 기록 조회(관리자용)", description = "유저의 아이디를 기반으로 참여한 설문 기록을 조회한다")
-    @GetMapping("participating/{memberId}")
-    public List<ParticipatedSurvey> getParticipatedSurveyList(@PathVariable Long memberId) {
-        return facadeRolePattern.getParticipatedSurveyList(memberId);
-    }
 
     @Operation(summary = "유저 설문조사 조회(관리자용)", description = "유저의 닉네임을 기반으로 ")
     @GetMapping("participating")
     public List<Survey> getParticipatedSurveyList(@RequestParam String nickname) {
         return facadeRolePattern.getMemberSurveyRecord(nickname);
+    }
+
+    @Operation(summary = "회원 정보 조회(관리자용)", description = "닉네임 기반 회원정보 조회")
+    @GetMapping("{nickname}")
+    public Member getMemberInfoByNickname(@PathVariable String nickname) {
+        return facadeRolePattern.getMemberInfoByNickname(nickname);
     }
 }
